requirements.txt
fastapi
uvicorn==0.30.1
tensorflow==2.16.1
biopython==1.83
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
python-multipart==0.0.9
mysql-connector-python==8.0.33
apache-airflow==2.7.0
passlib
python-jose